Pull runtime dep checks into their own module
